Game folder is uploaded to a server
The uploading is still "dumb", it is triggered by a hard
coded line in the beginning of the editor.py -file. Still it does
seem to work correctly, as all required files are uploaded,
including files in subfolders. NOTE that .html files should be
added to the excluded files later.

Still missing tests and comments, those must be added before this
feature can be accepted into the develop and main branches.