Catching missing attribute errors on startup
When starting the editor we get a lot of errors about
missing attributes. I just went around this by putting the
lines inside try-excepts. I don't know if this is a good idea
or not, as I don't really understand the errors yet, but it
seems to be working alright.