Improvements to template jsons
The template jsons are based off of the latkazombit jsons, but
they've simply had a bunch of rooms and objects removed. Honestly
it's a really shitty solution at this point, since it's just strange
to start a new adventure and already have some rooms, objects, and
a start and end screen - but what can you do.

We still have to figure out how to make the things like start menus,
etc editable..