Added debug logging to classes
Each function should now have a print in the beginning of it.
Each module has a DEBUG flag, that can be set to False to turn
this printing off.
Classes that use inheritance extract the name of the calling
class from the self-variable, and also print the actual location
of the method called.